feat: add base file to share functionality beetween buttons components
